resolve-2: imaginary literals actually work; fst.fx complex demo via 1.f + 1.fi
The lexer desugared N.fi/N.i/N.hi into complex(0, N) with an INT zero
(LitInt(0)), so 1.fi typed as complex(int, float) and no constructor
matched -- one of the original reasons the 2021 fst.fx demo was fenced
(FB-007 entry updated). The zero real part is now a float literal of the
same width as the imaginary part.

Also (Vadim): Complex.fx scalar-left variants regrouped next to their ops
(two stray 't2 from the reverted mixed-type experiment normalized back to
't -- mixed-type operator bodies are exactly FB-007/S1, session-2 scope);
fst.fx demo constructs the value as a literal expression (1.f + 1.fi) and
prints mathematically correct output now; print(int64/uint64) printf casts
fixed (C warnings). Golden 207: candidate order follows the new
declaration order.

Verified: test_all 135/135, fxtest all, determinism, sanitize,
vision_classify typechecks, bootstrap fixpoint holds (only Lexer.c).

+++ lib/Builtins.fx
@@ -1000,8 +1000,8 @@ fun print(a: 't) = print_string(string(a))
 @nothrow fun print(a: int16): void = @ccode { printf("%d", (int)a) }
 @nothrow fun print(a: uint32): void = @ccode { printf("%u", a) }
 @nothrow fun print(a: int32): void = @ccode { printf("%d", a) }
-@nothrow fun print(a: uint64): void = @ccode { printf("%llu", a) }
-@nothrow fun print(a: int64): void = @ccode { printf("%lld", a) }
+@nothrow fun print(a: uint64): void = @ccode { printf("%llu", (unsigned long long)a) }
+@nothrow fun print(a: int64): void = @ccode { printf("%lld", (long long)a) }
 fun print(a: long): void = @ccode {
     fx_cstr_t str;
     int fx_status = fx_ltoa_ascii(a, 'd', false, &str);
